+++ Classes/Utility/GeneralUtility.php
@@ -1,15 +1,25 @@
 <?php
+
 namespace SourceBroker\Ip2geo\Utility;
 
+use Exception;
+
+/**
+ * Class GeneralUtility
+ * @package SourceBroker\Ip2geo\Utility
+ */
 class GeneralUtility
 {
-    public static function getHash() {
+    /**
+     * @return bool|string
+     * @throws Exception
+     */
+    public static function getHash()
+    {
         $encryptionKey = $GLOBALS['TYPO3_CONF_VARS']['SYS']['encryptionKey'];
-
         if (!$encryptionKey) {
             throw new Exception('EncryptionKey not found.');
         }
-
-        return substr($encryptionKey,25,10);
+        return substr($encryptionKey, 25, 10);
     }
 }